requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cssselect==1.2.0
cycler==0.11.0
dnspython==2.3.0
filelock==3.12.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
html5lib==1.1
huggingface-hub==0.14.1
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
loguru==0.7.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.0
multidict==6.0.4
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.7
packaging==23.0
Pillow==9.4.0
pinecone-client==2.2.1
protobuf==4.22.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
pyquery==2.0.0
pyrouge==0.1.3
pysummarization==1.1.9
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.5.5
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
soupsieve==2.4.1
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tokenizers==0.13.3
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.5.0
urllib3==1.26.14
webencodings==0.5.1
Werkzeug==2.3.3
win32-setctime==1.1.0
yarl==1.9.2